Restructure document management components
